+++ src/runtime/Native/ITypeOffsets.cs
@@ -22,6 +22,7 @@ namespace Python.Runtime.Native
         int nb_true_divide { get; }
         int nb_and { get; }
         int nb_int { get; }
+        int nb_float { get; }
         int nb_or { get; }
         int nb_xor { get; }
         int nb_lshift { get; }
@@ -30,6 +31,7 @@ namespace Python.Runtime.Native
         int nb_invert { get; }
         int nb_inplace_add { get; }
         int nb_inplace_subtract { get; }
+        int nb_index { get; }
         int ob_size { get; }
         int ob_type { get; }
         int qualname { get; }