+++ tests/test_enum.py
@@ -15,7 +15,6 @@ def test_enum_standard_attrs():
     assert DayOfWeek.__name__ == 'DayOfWeek'
     assert DayOfWeek.__module__ == 'System'
     assert isinstance(DayOfWeek.__dict__, DictProxyType)
-    assert DayOfWeek.__doc__ is None
 
 
 def test_enum_get_member():
@@ -139,7 +138,7 @@ def test_enum_undefined_value():
     # This should fail because our test enum doesn't have it.
     with pytest.raises(ValueError):
         Test.FieldTest().EnumField = Test.ShortEnum(20)
-    
+
     # explicitly permit undefined values
     Test.FieldTest().EnumField = Test.ShortEnum(20, True)
 
@@ -157,6 +156,6 @@ def test_enum_conversion():
 
     with pytest.raises(TypeError):
         Test.FieldTest().EnumField = "str"
-    
+
     with pytest.raises(TypeError):
         Test.FieldTest().EnumField = 1